Detach Aggregates, AggregateStatistic and SortOrder Types into aggregates file
